@@ -5,19 +5,25 @@
     "dependencies": {
         "@dnd-kit/core": "^6.1.0",
         "@dnd-kit/sortable": "^8.0.0",
+        "@dnd-kit/utilities": "^3.2.2",
         "@emotion/react": "^11.11.4",
         "@emotion/styled": "^11.11.5",
         "@mui/icons-material": "^5.15.20",
         "@mui/joy": "^5.0.0-beta.48",
+        "@mui/lab": "^5.0.0-alpha.173",
         "@mui/material": "^5.15.20",
+        "@react-pdf/renderer": "^3.4.4",
         "axios": "^1.7.2",
         "js-cookie": "^3.0.5",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.3.1",
+        "react-chatbotify": "^1.7.0",
+        "react-dom": "^18.3.1",
+        "react-pdf": "^9.1.0",
         "react-router-dom": "^6.23.1",
         "react-scripts": "5.0.1",
         "react-select": "^5.8.0",
-        "recharts": "^2.12.7",
+        "react-to-pdf": "^1.0.1",
+        "recharts": "^2.13.0-alpha.4",
         "web-vitals": "^2.1.4"
     },
     "scripts": {
@@ -43,5 +49,9 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
+    },
+    "devDependencies": {
+        "@ag-media/react-pdf-table": "^1.0.1",
+        "@types/cypress": "^1.1.3"
     }
 }
