@@ -6,7 +6,7 @@ from datetime import timezone
 from typing import Literal, get_args
 Data_Type = Literal["float", "int"]
 Pillar = Literal["E", "S", "G"]
-
+Category = Literal["E", "S", "G"]
       #to do - change primary key from id to name - maybe dont need
 class Indicators(Base):
     __tablename__ = 'Indicators'
@@ -35,7 +35,13 @@ class Metrics(Base):
     id: Mapped[int] = mapped_column(primary_key=True, unique=True, nullable=False)
     name: Mapped[str] = mapped_column(String(100), unique=False, nullable=False)
     # description: Mapped[str] = mapped_column(String(1000))
-
+    category: Mapped[Category] = mapped_column(Enum(
+      *get_args(Category),
+      name="category",
+      create_constraint=True,
+      validate_strings=True,
+    ))
+    
 class MetricIndicators(Base):
     __tablename__ = 'MetricIndicators'
     
@@ -43,3 +49,15 @@ class MetricIndicators(Base):
     metric_id: Mapped[int] = mapped_column(primary_key=False, unique=False, nullable=False)
     indicator_name: Mapped[str] = mapped_column(String(100), primary_key=False, unique=False, nullable=False)
     indicator_id: Mapped[int] = mapped_column(primary_key=False, unique=False)
+    weighting: Mapped[float] = mapped_column(primary_key=False, unique=False)
+
+class CustomMetricIndicators(Base):
+    __tablename__ = 'CustomMetricIndicators'
+    
+    id: Mapped[int] = mapped_column(primary_key=True, unique=True, nullable=False)
+    metric_id: Mapped[int] = mapped_column(primary_key=False, unique=False, nullable=False)
+    framework_id: Mapped[int] = mapped_column(primary_key=False, unique=False, nullable=False)
+    indicator_name: Mapped[str] = mapped_column(String(100), primary_key=False, unique=False, nullable=False)
+    indicator_id: Mapped[int] = mapped_column(primary_key=False, unique=False)
+    user_id: Mapped[int] = mapped_column(primary_key=False, unique=False, nullable=False)
+    weighting: Mapped[float] = mapped_column(primary_key=False, unique=False, nullable=False)
